Use a real docker: ImageSource to make image data available for verification
Drop most of the fake sigtypes.* implementations.  The local unparsedImage
still exists, but only to override .Signatures() to the one signature we
want to use.  Otherwise, we use a real UnparsedImage based on a real
ImageSource; in particular, this can fetch the true manifest blob
which is no longer available from the etcd-stored API object.

o.ExpectedIdentity is now handled by constructing an appropriate policy,
instead of overriding the image location and blindly hoping that the
policy will ask for a location-based match.

Uses ClientConfig.BearerToken to authenticate to the registry;
presumably a bit more setup like this will be needed (e.g. to disable
TLS if necessary), but I can’t find a template to copy that from.